Add more modules to make it work properly
