+++ src/index.html
@@ -13,7 +13,7 @@
         </div>
         <div class="dflex spacebetween">
             <input type="text" class="textinput" placeholder="Add to your list">
-            <div class="dflex"><img class="icon" src="./img/entericon.png" alt="enter"></div>
+            <div class="dflex"><img class="icon enterbutton" src="./img/entericon.png" alt="enter"></div>
         </div>
         <div class="listelementcontainer dflex flexcol">
