+++ src/index.js
@@ -13,14 +13,22 @@ import storageAvailable from './modules/isStorageValid.js';
 let arraylist = [];
 const isStorage = storageAvailable('localStorage');
 
+const addelement = (event) => {
+  createlistelement(inputElement.value, arraylist);
+  localStorage.setItem('ToDoList', JSON.stringify(arraylist));
+  iteratearray(arraylist);
+}
+
 const inputElement = document.querySelector('.textinput');
+const enterbutton = document.querySelector('.enterbutton');
 inputElement.addEventListener('keyup', (event) => {
-  if (event.keyCode === 13) { // 13 is the keycode for 'Enter' key
-    event.preventDefault();
-    createlistelement(inputElement.value, arraylist);
-    localStorage.setItem('ToDoList', JSON.stringify(arraylist));
-    iteratearray(arraylist);
-  }
+  event.preventDefault();
+  if (event.keyCode === 13 && inputElement.value != '') addelement(event);
+});
+
+enterbutton.addEventListener('click', (event) => {
+  event.preventDefault();
+  if (inputElement.value != '') addelement(event);
 });
 
 document.addEventListener('DOMContentLoaded', () => {
