+++ src/modules/iteratearray.js
@@ -1,45 +1,6 @@
+import editelement from './editelement.js';
+
 const listelementcontainer = document.querySelector('.listelementcontainer');
-
-const switchelement = (inputelement, elementcontainer, arr, i) => {
-  const newlistelement = document.createElement('div');
-  newlistelement.classList.add('listelement');
-  newlistelement.innerText = inputelement.value;
-  inputelement.parentNode.replaceChild(newlistelement, inputelement);
-  newlistelement.addEventListener('dblclick', () => {
-    editelement(newlistelement, elementcontainer, arr, i);
-  });
-}
-
-const editelement = (listelement, elementcontainer, arr, i) => {
-  const threedots = elementcontainer.querySelector('.threedotsicon');
-  const deletebutton = elementcontainer.querySelector('.deleteicon');
-  deletebutton.classList.remove('dnone');
-  threedots.classList.add('dnone');
-  const newinput = document.createElement('input');
-  newinput.type = 'text';
-  newinput.classList.add('yellowbg');
-  newinput.value = listelement.innerText;
-  listelement.parentNode.replaceChild(newinput, listelement);
-  newinput.focus();
-  newinput.select();
-  deletebutton.addEventListener('click', () => {
-    arr.splice(i,1);
-    for (let j = i; j < arr.length; j+=1) {
-      arr[j].index = j + 1;
-      console.log('arr[j] is: ', arr[j]);
-    }
-    localStorage.setItem('ToDoList', JSON.stringify(arr));
-    console.log('arr: ', arr);
-    iteratearray(arr);
-  });
-  newinput.addEventListener('blur', () => {
-    setTimeout(() => {
-      deletebutton.classList.add('dnone');
-      threedots.classList.remove('dnone');
-      switchelement(newinput, elementcontainer);
-    }, 100);
-  });
-}
 
 const iteratearray = (arr) => {
   listelementcontainer.innerHTML = ' ';
